#430: city culture was OVERFLOWING int32 -- widen it, and delete the guards that hid it
Step 2 of the width work. The enumeration found one field that genuinely needed
it, and the evidence was already sitting in the getters.

CvCity::m_aiCulture accumulates getCommerceRateTimes100(CULTURE) every turn and
NEVER decays -- pure accumulation at x100. It wrapped, and the wrap was papered
over rather than fixed:

    getCulture:          m_aiCulture[i] < 0 ? MAX_INT / 100 : m_aiCulture[i] / 100
    getCultureTimes100:  m_aiCulture[i] < 0 ? MAX_INT       : m_aiCulture[i]
    setCulture:          iNewValue > MAX_INT / 100 ? MAX_INT : 100 * iNewValue

Those `< 0` tests are OVERFLOW DETECTION -- they exist because the value goes
negative in real games -- and saturating to MAX_INT silently corrupts every
consumer: culture percent, cultural ownership, the level thresholds. Same shape
as a self-heal being the fossil of a missing emit: the guard is the fossil of
the overflow. All three are deleted with the overflow they existed for.

The seam was also provable rather than inferred: CvPlayer::m_iCulture is ALREADY
int64_t. One quantity, int64 at player scope and int32 at city scope.

SAVE MIGRATION (save.md 8, its first real use): int -> int64_t is a type-code
change, so the value moves to a new tag. The wide tag reads into the member; the
old 32-bit tag reads into a scratch array and seeds it when the wide one was
absent -- which is what lands an old save on its real value rather than a
default. Safe both directions: on a new save the legacy tag is absent and its
scratch stays 0, and seeding 0 over 0 is a no-op. NO savemigration.txt entry --
the old tag is still genuinely READ, so it is not an orphan yet.

Consumers that would have relocated the truncation are widened with it (the
acquireCity scratch array, the espionage culture-insert amounts, the city-side
culture locals) -- widening storage while a consumer still assigned to int would
have moved the bug, not fixed it.

Checked and NOT widened, with reasons: unit experience (x100 but XP is small),
great-people progress and stored food (reset at their thresholds), team war
weariness (x100 but decays every turn, so bounded), plot culture (x100 but
proportionally decays to an equilibrium). Player gold/culture/expense were
already int64_t.

Assert build: no error names any culture symbol; conversion-error count back to
the pre-existing baseline; no narrowing warnings.

+++ Sources/Engine/CvCity.h
@@ -1068,16 +1068,16 @@ public:
 	int getDomainProductionModifier(DomainTypes eIndex) const;
 	void changeDomainProductionModifier(DomainTypes eIndex, int iChange);
 
-	int getCulture(PlayerTypes eIndex) const;
-	int getCultureTimes100(PlayerTypes eIndex) const;
-	int countTotalCultureTimes100() const;
+	int64_t getCulture(PlayerTypes eIndex) const;
+	int64_t getCultureTimes100(PlayerTypes eIndex) const;
+	int64_t countTotalCultureTimes100() const;
 	PlayerTypes findHighestCulture() const;
 	int calculateCulturePercent(PlayerTypes eIndex) const;
 	int calculateTeamCulturePercent(TeamTypes eIndex) const;
-	void setCulture(PlayerTypes eIndex, int iNewValue, bool bPlots, bool bUpdatePlotGroups, bool bNationalSet = false);
-	void setCultureTimes100(PlayerTypes eIndex, int iNewValue, bool bPlots, bool bUpdatePlotGroups, bool bNationalSet = false);
-	void changeCulture(PlayerTypes eIndex, int iChange, bool bPlots, bool bUpdatePlotGroups);
-	void changeCultureTimes100(PlayerTypes eIndex, int iChange, bool bPlots, bool bUpdatePlotGroups);
+	void setCulture(PlayerTypes eIndex, int64_t iNewValue, bool bPlots, bool bUpdatePlotGroups, bool bNationalSet = false);
+	void setCultureTimes100(PlayerTypes eIndex, int64_t iNewValue, bool bPlots, bool bUpdatePlotGroups, bool bNationalSet = false);
+	void changeCulture(PlayerTypes eIndex, int64_t iChange, bool bPlots, bool bUpdatePlotGroups);
+	void changeCultureTimes100(PlayerTypes eIndex, int64_t iChange, bool bPlots, bool bUpdatePlotGroups);
 
 	int getNumRevolts(PlayerTypes eIndex) const;
 	void changeNumRevolts(PlayerTypes eIndex, int iChange);
@@ -1770,7 +1770,7 @@ protected:
 	int* m_aiProductionToCommerceModifier;
 	int* m_aiCommerceRateModifier;
 	int* m_aiDomainProductionModifier;
-	int* m_aiCulture;
+	int64_t* m_aiCulture;   // per-player culture, x100 and NEVER decaying -- an AMOUNT, so 64-bit
 	int* m_aiNumRevolts;
 
 	bool* m_abEverOwned;
